SRV-1379: Fixed inconsistency with string-deserialization of response
Some gateway-API methods return pure strings as return value (e.g.
addNode returns the node id, e.g. 5:10) that is transferred as
json-object in case of the REST-gateway-client - i.e. quotes are added
to the string. However, when deserialized on client-side, those quotes
are not removed, which is inconsistent!

SRV-1379 (Create testing fragment for gateway projects)