SRV-1939: EntityProxyTable: cache individual rows instead of entire chunks
Not needed right now but makes the caching more robust when the chunks
are accessed in an 'inconsistent' manner, i.e. when the assumption that
"from" is always a multiple of (the same) "count" is not fulfilled.

SRV-1939 (Support table port view of new DB nodes in remote workflow
editor)