WEBP-269: WEBP-269: Support for wizard pages in loops added
I.e. a wizard page is in a loop it is asssumed that this loop will
continue and the uppler limit of node executions will be reported
respectively. Until the execution leaves the loop scope.
Multiple nested loops not supported, yet.

WEBP-269 (Return upper limit for progress between WebPortal pages at the execution-statistics endpoint)