SRV-2696: Exception mechanism added to deal with new endpoints
Such that an explaining exception is thrown if client tries to connect a
new endpoint not existing on the server.
New endpoints now have a since-version and a 'missing message' and
client version is compared to server version (which is now known to the
gateway rest client implementations).
Reason: to still allow new clients to connect to older server-versions
but with reduced features.

SRV-2696 (Enable dynamic port configuration in remote workflow editor)