WEBP-278: New hasNextPage property in wizard page response
... that indicates whether there is a next page.
Special case handling if the very last node is a wizard page and the
workflow has a report.
Wizard execution state logic changed, too - if a wizard page is
represented by the very last node in a branch, it will never have the
state 'interaction_required'.

WEBP-278 (There should be no Next button after last Component in the Workflow)