NXT-237: Fixed patch operation for changes to node id
E.g. if the index of a node id, e.g., for a connection (source) is
changed the respective patch operation (to replace the entire node id
entity) is created now. Previously the change has been omitted.

Fix is to treat all id-objects (for nodes, connections and annotations)
as atomic values. In the process of it the PatchChangeProcessor needed
some adoptions, too.

NXT-237 (Allow user to replace a connection in remote workflow editor)