NXT-237: WorkflowManager transformations moved
Which are used by the event service to test that the respective events
are triggered for certain transformations.
But now also used (thus the move) to test the correctness of the patches
created for the remote workflow editor (java-ui).

NXT-237 (Allow user to replace a connection in remote workflow editor)