NXT-264: Info about allowed actions added to workflow
I.e. information whether certain actions (such as execute, reset or
cancel) can be performed on individual nodes or the entire workflow itself.
Required to enable/disable the respective buttons.

The inclusion of the info is optional and can be controlled by a
parameter.

NXT-264 (Gateway API: Execution, Cancelation, Reset)