NXT-488: Determiniation of executing successors and ...
... executable predecessors refactored at the component and metanode
borders (i.e. connections leaving/entering compnents/metanodes).
Instead of adding the nodes of the upper level to the queue, the nodes
at the component/metanode 'borders' are added.

Test extended to include the special case of double-nested metanode
directly connected to its parent metanode (which was the original
cause for this issue).

NXT-488 (Metanode fails to open)