Test snapshot fixed
Some workflow transformations (WorkflowChangedEventsTest) don't
produce a deterministic number of events (which is fine in
general - but tricky for testing; at least it's either 1 or 2 events).
For some transformations we just test both cases. Though, for the
'executed_node'-transformation the '2-event-case' was compared against a
wrong snapshot - the port-object version of the flow variable port (port
0) is not expected to change.