NXT-965: When transferring stack elements, add to head
In this implementation the top of the stack is the head of the
underlying Deque. Thus, when adding an element to the top of the
stack, we have to add it to the head of the Deque, instead of
add the end as done previously.

NXT-965 (Undo not working when deleting a single element)