NXT-1130: Don't fail if 'DependentNodeProperties' are 'out of sync'
... with the nodes actually contained in the workflow. I.e. in case the
DependentNodeProperties don't contain a node-id which is already in the
workflow, it just returns 'false' for all respective properties. By that
we'll could have an inconsistent intermediate state - but that's guaranteed to
be only for a short instant. Because the event which actually, e.g.,
added or removed a node to/from the workflow, just didn't arrive, yet
(but will shortly after).
Tried to come up with a test - but wasn't able to provoke the problem
with a unit test.

NXT-1130 (Collapsing a single node into a component fails under some circumstances)