NXT-1206: Make test more robust
Don't let the node recommendations test check for the exact number of
nodes returned because those can vary depending on which nodes are
available in the test setup. And the touched test is not about testing
the correctness of the node recommendation engine but just of the node
recommendation endpoint.

NXT-1206 (Gateway API: Suggested nodes for successor nodes)