NXT-1378: Sort of 'namespacing' introduced for the EntityBuilderUtil
Methods are now organised in extra 'Builder'-classes in order to
separate the builder-methods which are (mostly) independent from each
other. Should, first, simplify the usage of the methods (by first narrowing down
the namespace to use the method from) and, second, make the maintainance
a bit easier.

NXT-1378 (Gateway API: local workspace browsing)