NXT-1402: Sort search results based on collection and name
If all nodes are included in a search result the nodes that are part of
the configured collection should still be the first nodes listed.
Also, we sort by the name as a last resort to keep the order of nodes
that have the same score and weight consisted in searches with or without
"includeAll".

NXT-1402 (Node repo based on a set of nodes or all nodes)