NXT-1448: NodeFactory-info moved one level up
From the NodeTemplate-entity (mainly/only used within the node
repository) to its parent, NativeNodeInvariants-entity (only used within
a workflow to provide a node's icon etc.).
Why? Because by the node-factory-info is required to fetch the node's
node description (to identify the node) which, so far, was only
necessary for nodes in the node repository. Now we need to do that for
nodes in the workflow, too.

NXT-1448 (Display node description in info panel for selected node inside workflow)