NXT-1432: 'ancestorItemIds' added to workflow project origin
I.e. added to the workflow project origin-entity shipped to the FE.
The ancestor-item-ids are provided by a Space (and here implemented by
the LocalWorkspace). As a result, the application-service and the
event-service (both are creating workflow-project-entities) now require
SpaceProviders as a dependency.
Also: bit of refactoring around service-tests as a result of the newly
added service dependency.

NXT-1432 (Provide all ancestor ids of open workflows)