NXT-1504: 'active workflow' removed from workflow project entity (app state)
... and replaced by just the active workflow id.
The actual workflow is now fetched by the FE separately.
Makes more sense and solves this bug as a positive side effect because a
workflow (snapshot) is now created only at one single 'entry point' (the
get-workflow endpoint) and _not_ when creating the application state.

The actual issue:
When creating the workflow as part of the application state, the
WorkflowBuildContext wasn't initialized correctly, i.e. the
undo/redo flags weren't set. As a result, when updating the app state,
there were workflow differences, resulting in the creation of a new
snapshot (and snapshot-id) which the FE didn't properly update on its
end (presumably).

NXT-1504 (Closing (and saving) workflows leads to an impaired state)