NXT-1537: Improve the `LocalWorkspace.moveItem(...)` method.
Instead of using `Files.getFileStore(...)` to probe for different
file systems when attempting to move files, this introduces a
try-catch-mechanism, since the first approach is not supported on
every OS and in every environment. It depends on the `FileStore`
implementation used by the JVM and how it resolves file locations,
etc.

NXT-1537 (Cannot save a newly created Hub workspace locally)