NXT-1004: Avoid serialization of any fields in the implementations
... of entity-interfaces. Make sure only the methods in the
interfaces are being serialized.

Because the implementations are now record classes which in turn has
public fields which are otherwise considered for the serialization. And
in case of, e.g., a field named _private and a method named isPrivate,
those are serialized twice (with '_private' and 'private').

NXT-1004 (Use Java 17 language features)