NXT-1734: portObjectVersion renamed to portContentVersion and semantics changed
It's now the version of either the port data (in case the node is
executed) or the port spec (in case the node is not executed).
Can still be null/absent in case there is no port object spec (e.g.
Transpose-node).

NXT-1734 (Communicate port-spec-updates and null-port-specs to the frontend)