NXT-1847: DefaultServiceContext introduced
Which allows the default service implementations to be limited to a
single workflow project only.
If a specific workflow project id is set with the default-service-context,
access to other projects will fail and the app state will only
contain a single project.

NXT-1847 (WebSocket server for the KNIME AP Executor)