NXT-2032: Provide no-op initial selection supplier
... which just returns an empty initial selection in case of embedded
port view. Because the table port view is currently configured in a way
that it supports selection (which is does in case of the detached port
view but not for the embedded port view, yet).

Will be re-visited as part of NXT-2042.

NXT-2032 (Enable functionality for hiliting / selection for detached(!) table port view)