NXT-2213: Only determine ancestorItemIds for local projects
To avoid unnecessary requests in case of hub projects every time the app
state is updated. And hub projects don't need it because they are not
associated with space items.

NXT-2213 (App state not properly not updated if there is open hub project which has been deleted on the hub)