NXT-2179: Properly remove all workflow service projects on first perspective switch
... from modern to classic UI.
Because in the classic UI those are tracked by another data structure
which is not filled at that point which otherwise leads to an inconsistency
(because LocalWorkflowBackends are cached but not registered with the
classic UI's ProjectWorkflowMap).
Hence, we just remove all of them on the first perspective switch by
calling the LocalWorkflowBackend to do so (via a callback to inverse the
dependency).

NXT-2179 (Workflows already loaded via 'Call Workflow (Service)' aren't taken into account)