UIEXT-932: Remove js-core dependency from org.knime.gateway.testing.helper
Because it's a plugin which is also used by knime-com-gateway and hence
is also included in the 'non-test-profile' to be available on a
respective update site. As result it must not have a dependency to
knime-js-core in order to avoid cycles (knime-js-core has a dependency
to knime-gateway).

UIEXT-932 (Move gateway entities from knime-core-ui to knime-gateway)