AP-21872: Fully reset metanodes on "Reset All" action
This commit adds extra resetting of metanodes in the workflow. This will have the effect of
fully resetting the metanode (all contained nodes). Before, the "Reset All" routine would only
reset nodes in the metanode that are successors to source nodes.
AP-21872 (Reset All button doesn't function properly for metanodes)