NXT-2435: Avoid deadlocks between workflow-command handler threads and core workflow change listeners.
The following deadlock was identified. With the classic UI initialized,
1. Trigger translation of many workflow elements including connections with bendpoints.
2. Then, very quickly, trigger translation of another workflow element (e.g., an annotation)

This caused the following thread state:

(a) The workflow-command handler thread (named e.g., `comm-pool-thread-<K>`), in sequence
  (a.1) acquires and holds a WorkflowLock (by means of `AbstractWorkflowCommand#executeWithLockedWorkflow`)
  (a.2) makes changes to the workflow
  (a.3) these workflow changes *synchronously* trigger event listeners of the classic UI, in particular
    `ConnectionContainerEditPart#connectionUIInformationChanged`
  (a.4) this listener wants to execute something synchronously on the SWT Display thread. To do so,
    it must acquire a lock on `org.eclipse.swt.widgets.Synchronizer`.

(b) The `main` thread is associated with the Eclipse RCP Workbench. In particular, the `PartRenderingEngine`
  will take care of re-rendering Edit Parts in the classic UI when there have been changes to the model.
  (b.1) Any such update is performed by invoking a runnable on the Display thread. Due to this, a lock on
    `Synchronizer` is obtained.
  (b.2) A change listener might try to acquire a WorkflowLock (`AbstractEditPart#refreshChildren`).

We now have a deadlock where (a) is blocked in (a.4) by waiting for a resource currently held by (b.1),
while (b) is blocked in (b.2) by waiting for the WorkflowLock held by (a.1).

This deadlock was introduced by commit f93f90cb in knime-workbench, which changed the classic UI event
handler of connections to be dispatched on the Display thread. With the addition of the Modern UI,
the caller is no longer necessarily the Display thread. This dispatch is synchronous to maintain the
synchronicity of any other callers. See usages of `ConnectionContainer#setUIInfo`.

This commit mitigates this problem by eliminating the dependence of (a) on the Display thread
(i.e. `Synchronizer` lock). This is already the case for many workflow modifications but cannot be assumed
in general.
NXT-2435 (Deadlock in modern UI when dragging annotations  (in 'hybrid' mode))