NXT-3169: NXT-3169: Add copy and delete directory mechanism when a rename fails
Rename can fail in windows when a subdirectory of a workflow is opened in the filesystem. This copy and delete additional step is similar to the existing approach in classic UI. Look class GlobalRenameAction.java in Workbench for more details.

NXT-3169 (Cannot rename workflow if workflow's data folder is open in Windows file explorer)
NXT-3169 (Cannot rename workflow if workflow's data folder is open in Windows file explorer)