NXT-3125: Remove term "workspace" where irrelevant
Sometimes we would refer to a `org.knime.impl.webui.spaces.Space` informally as "workspace" which is a bit noisy. If anything, this is only ever relevant to the "local" space. Even then, to be consistent with `HubSpace` and `ServerSpace`, it should be called `LocalSpace`.

The "workspace" is a term from Eclipse RCP, which should only be used if we are actually dealing with such a thing (and not a Space in the MUI sense). Further, most usages of the local space are not particularly concerned whether it is is an RCP workspace, hence the rename `LocalWorkspace` -> `LocalSpace`.