HUB-9174: Supply projectId when adding/removing event listeners
... for event sources.
It's the project-id from the DefaultServiceContext, which is only
available if service-call are associated with a particlar project.
The project-id is used to validate project-id parameters and, e.g., to
make sure only a space provider for a particular project is used (in the
respective event source).

HUB-9174 (Display space explorer sidebar in browser)