HUB-9174: Simplify
AppStateEntityFactory only needs the right `SpaceProviders` instance. Previously, we would pass in `SpaceProvidersManager` plus a parameter to select from it. But the `AppStateEntityFactory` need not (and must not) be concerned with this, it only needs a (right) `SpaceProviders` instance. This also makes the method signatures simpler.

Similarly for WorkflowCommands.

(HUB-9174: Display Space Explorer Sidebar in Browser)