NXT-135: De-duplicate wfm loading logic; Make wfm loading progress reporting a DesktopAPI dependency.
This was motivated by considering the necessity of a specific WorkflowManagerLoader that provides only the current version and fails if another is requested.

Testing code can now instead use a separate builder method as suggested.

 In "production" code, we create and open Projects via two places: (1) DesktopAPI>OpenProject and (2) AppStatePersistor>deserialize. The latter case initially featured such a WorkflowManagerLoader providing only a current version.

This is not necessary because, when deserializing open Projects, there is no inherent reason why the Project should not be able to provide versions as well. The constraint only came from that AppStatePersistor>deserialize had its own (duplicated) logic to load a single (!) WorkflowManager instance and then use this loaded instance to create a Project. However, from the perspective of gateway code, the restored Project might just as well also provide versions. Whether versions can be provided is rather a property of the Space(Provider).

So, this commit instead adjusts LocalSpace#toLocalAbsolutePath to fail if a version other than current-state is requested because the Space(Provider) simply does not support that.

This then means that there is no difference between how a WorkflowManager is loaded and a Project is created between (1) OpenProject and (2) AppStatePersistor. In both cases, we can now configure a project with a `WorkflowManagerLoader` that specifies how a WorkflowManager instance is loaded from the `Origin` of the Project. This eliminates the code duplication we previously had between (1) and (2).

Regarding production code, the WorkflowManagerLoader could simply be implemented as previously by `DesktopAPUtil.consumerWithProgress(..., DesktopAPUtil.fetchAndLoadWorkflowWithTask(...))`. The outer method provides an `IProgressMonitor` to the inner method to which it reports progress or exceptions. Concretely, up to now, this shows SWT progress bars or modals.

The problem with this, though, is that this does not work in headless unit tests (or generally any other cases in which you do not want to use that kind of progress reporting). This also seems to have been the motivation why `OpenProject` was previously structured as

```
// API entry point
static void openProject(...) {
  DesktopAPUtil.consumerWithProgress(..., openProjectWithProgress(...)
}

// only called from #openProject and tests
static void openProjectWithProgress(IProgressMonitor, ...) {
  // actually load and open the project
}
```
such that `openProjectWithProgress` can be called from `OpenProjectTest` with a `NullProgressMonitor` argument.

Instead of exposing methods to make it possible to configure their dependencies, we can also use the "service dependency" (or DesktopAPI dependency) concept which we already employ in many places (and in that specific test, too). This commit adds a DesktopAPI dependency `ProgressReporter` that can be implemented by `DesktopAPUtil#consumerWithProgress` (which is essentially `busyCursorWhile` plus catching) providing SWT modals. This makes it possible to test the code without making amends and allows some simplifications. Further, this factors out the SWT/workbench dependency and makes it ready to replace it: The `ProgressReporter` can readily become a gateway service dependency and, instead of showing a SWT progress modal, send progress info to the frontend, which may display a modern progress indicator which can also be used in a browser context.  (there's still more SWT dependency deeper in the callstack in `LoadWorkflowRunnable` though, but it's a start).

NXT-135 Restore Version