NXT-3288: Remove item from MostRecentlyUsedProjects state immediately; never abort shutdown due to throws in app state saving
The issue was that the item has been removed from the Space, but not from the list of MostRecentlyUsedProjects, causing a NoSuchElementException when trying to serialise it.

This commit resolves it by having MostRecentlyUsedProjects attach a listener to LocalSpace that updates the MRU state as soon as a space item is deleted. Previously, the list would be updated ad-hoc on frontend request. This bug shows that this is not sufficient (a crucial step to reproduce is to not switch back to the Recent page, hence updating the state). Because the state is now updated immediately (no matter the user/frontend interactions), the ad-hoc update is no longer needed and the DesktopAPI method has been renamed accordingly.

(This could potentially be extended to other kinds of Spaces, too, but there we have no means to identify whether it wasn't deleted on the remote side without a request)

This commit also adds a try/catch to never abort shutdown on exceptions such as these.

NXT-3288 (Cannot close AP: AppStatePersistor throws NoSuchElementException)