NXT-135: Introduce WorkflowManagerCache as a service dependency.
(Note, in the previous commit `WorkflowManagerCache` was specific to a project, this now became `ProjectWfmCache`).

It is used to connect WorkflowMiddleware and ProjectManager/Project and their associated WorkflowManager instances. WorkflowMiddleware has to be notified when a (version) wfm is disposed s.t. it can clean up its associated state. Alternatives would have been to introduce a `Version` object (not really needed) or dispose via ProjectManager#disposeVersion (not very object-oriented).

Another motivation is that in a future multi-user setting, this cache can be specific to the user (though this could also be handled by making ProjectManager user-specific).

Version wfm dispose can be triggered not only through an explicit call (e.g. on "restore version", see DefaultWorkflowService#disposeVersion), but also on Project#dispose (e.g. on project close). To handle this cleanly, `Project`s are now backed by a `ProjectWfmCache` instance that is configured to notify the `WorkflowManagerCache` service dependency (see OpenProject#createProjectFromOrigin).

Further changes:
- Rename `Project#WorkflowManagerCache` to `ProjectWfmCache` and move to separate file.
- Further unify/streamline how `Project` instances are created to reduce ambiguity/duplication between AppStatePersistor, OpenProject and SaveProjectCopy. In all cases, `Project`s are created and loaded based on an `Origin`.
- Change Lazy#ifInitialized to Lazy#ifPresent to be consistent with Optional#ifPresent
- Use Project#updateOrigin instead of creating a new instance via builder to make sure that other properties (such as configured cache) stays intact.
- Remove Project#Builder#onDispose. The same calls are already being made by ProjectWfmCache#disposeWorkflowManager via CoreUtil#cancelAndCloseLoadedWorkflow.
- Add `Listeners` utility.
- Remove duplicate method LocalSpace#getUniqueProjectId

NXT-135 (Restore Version)