NXT-3674: Provide WorkflowManagerLoader with Projects in GatewayWebSocketServer Session(s) for GatewayWebsocketServer
(a) Remove duplication between dev and executor setup: GatewayWebSocketServer and GatewayDevServer were both doing essentially the same thing. This commit changes GatewayDevServerApplication to invoke GatewayWebSocketServer, which is now fully parameterised (`ServerContext`) to accommodate for both settings. GatewayDevServer is removed. Besides removing duplicated code, this allows locally running and testing the GatewayWebSocketServer class without a full executor/hub setup. This also makes it a bit easier to make the class unit-testable.

(b) For dev server, allow loading a project from a hub, identified by the provider-space-item-ID parameters. These were previously used to provide needed context for features as the space explorer. This comes practically for free with the previous changes which enabled creating a `Project` with an `Origin` and a `WorkflowManagerLoader`. Parameter `workflowDir` will now be optional, if present the local workflow is loaded instead as previously.

NXT-3674 (Load version WorkflowManager instances in GatewayWebsocketServer and GatewayDevServer)