NXT-3331: Fix ProjectManager.getProjectIds(ProjectConsumerType)
Which only differentiated between ui- and non-ui consumer type when
filtering which is wrong with the the introduction of the 'virtual
workflow' consumer type.
Also: no need to have a count for each project-consumer-type since there
can actually only be one of each type per project.

NXT-3331 (Workflow can't be deleted when run as a 'workflow service' before)