NXT-4224: Adding workflow syncer and it's dependencies
Some considerations while iterating on the implementation:

Consideration 1:
To instantiate a WorkflowSyncer, we require a SpaceProvider. Previously, this would be done via getSpaceProvider, which does a lookup in SpaceProvidersManager.
However, it might be the case that SpaceProvidersManager is no yet populated. The assumption that this is the case rests on call order:
1. server initialisation creates empty SpaceProvidersManager instance
2. server init. creates WorkflowSyncerManager with SpaceProvidersManager as parameter. The parameter is used to create a WorkflowSyncer.
At this point, a call to WorkflowSyncerManager#getWorkflowSyncer would fail because there is not yet a SpaceProvider in SpaceProvidersManager
3. SpaceProvidersManager is populated
4. A call to WorkflowSyncerManager#getWorkflowSyncer is made, which calls the computeIfAbsent lambda, which accesses SpaceProvidersManager
In other words, SpaceProvidersManager has to be populated before we first access WorkflowSyncers, but this dependency is not visible. It is clearer to avoid the implicit instantiation in WorkflowSyncerManager#getWorkflowSyncer#computeIfAbsent and instead create the instance explicitly. Required parameters (SpaceProviders) can be given directly to the constructer, making the dependency obvious and impossible to miss.

Consideration 2:
Attach & dispose syncer instance cleanly --
A syncer ultimately is specific to a WorkflowManager instance because it listens to events emitted by a wfm. Indeed, currently, a Session corresponds to a Project corresponds to a WorkflowManager. However, the WorkflowManager instance is not guaranteed to be present for the entire session, see e.g. WorkflowService#disposeVersion. So we need to react to the wfm being disposed. This could be implemented by a Project#onWfmDispose callback (as in previous commits). But then in the callback implementation one would have to identify the syncer instance to call `dispose` on, i.e. have to know the project/session ID and then lookup the syncer in WorkflowSyncerManager. Attaching the syncer directly to the wfm via WorkflowResourceCache means (a) We do not need to add listener support in Project/ProjectWfmCache, (b) we do not need to register the on-dispose procedure of the syncer, complicated by a syncer only being optionally present, (c) we do not have to implement an access facility (previously the WorkflowSyncerManager/WorkflowSynerProvider service dependency) but can use existing API.

Note that two things have to be considered: (1) the syncer has to be attached to the wfm instance already loaded in GatewayJobPool (2) the syncer has to be attached to any wfm dynamically loaded in the future. For separation of concerns, (2) should not affect the WorkflowManagerLoader implementation (as in previous commits). (1) has to be done from the session initialisation. For (2) I can currently think of no better way than to expose a Project#onWfmLoad callback (as in previous commits).

NXT-4224 (Automatically / Manually sync when editing workflows in the browser)