fix implicit lifecycle assumptions; simplifications
To instantiate a WorkflowSyncer, we require a SpaceProvider. Previously, this would be done via getSpaceProvider, which does a lookup in SpaceProvidersManager.

However, it might be the case that SpaceProvidersManager is no yet populated. The assumption that this is the case rests on call order:
1. server initialisation creates empty SpaceProvidersManager instance
2. server init. creates WorkflowSyncerManager with SpaceProvidersManager as parameter. The parameter is used to create a WorkflowSyncer.

At this point, a call to WorkflowSyncerManager#getWorkflowSyncer would fail because there is not yet a SpaceProvider in SpaceProvidersManager

3. SpaceProvidersManager is populated

4. A call to WorkflowSyncerManager#getWorkflowSyncer is made, which calls the computeIfAbsent lambda, which accesses SpaceProvidersManager

In other words, SpaceProvidersManager has to be populated before we first access WorkflowSyncers, but this dependency is not visible. It is clearer to avoid the implicit instantiation in WorkflowSyncerManager#getWorkflowSyncer#computeIfAbsent and instead create the instance explicitly. Required parameters (SpaceProviders) can be given directly to the constructer, making the dependency obvious and impossible to miss.

Added benefit is that since the WorkflowSyncerManager no longer needs to be capable of creating WorkflowSyncer instances, it can be simplified to only a Map.